Moved SwitchTab from TabComponent to Subreddit
