+++ reactfe/src/RedditApp/BodyComponent/ContentComponent/Subreddit/SwitchTab/Home/index.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react'
-import Context from '../../../../../provider'
+import Context from '../../../../../../provider'
 import {
     Container,
     Alert,
